+++ innd/nc.c
@@ -746,8 +746,9 @@ NCproc(CHANNEL *cp)
 
   for ( ; ; ) {
     if (Tracing || cp->Tracing) {
-      syslog(L_TRACE, "%s cp->Start=%d cp->Next=%d bp->Used=%lu",
-        CHANname(cp), cp->Start, cp->Next, (unsigned long) bp->used);
+      syslog(L_TRACE, "%s cp->Start=%lu cp->Next=%lu bp->Used=%lu",
+        CHANname(cp), (unsigned long) cp->Start, (unsigned long) cp->Next,
+        (unsigned long) bp->used);
       if (bp->used > 15)
 	syslog(L_TRACE, "%s NCproc state=%d next \"%.15s\"", CHANname(cp),
 	  cp->State, &bp->data[cp->Next]);
@@ -889,8 +890,8 @@ NCproc(CHANNEL *cp)
       TMRstop(TMR_ARTPARSE);
       if (cp->State == CSgetbody || cp->State == CSgetheader ||
 	      cp->State == CSeatarticle) {
-	if (cp->Next - cp->Start > innconf->datamovethreshold ||
-	  (innconf->maxartsize > 0 && cp->Size > innconf->maxartsize)) {
+        if (cp->Next - cp->Start > (unsigned long) innconf->datamovethreshold
+            || (innconf->maxartsize > 0 && cp->Size > innconf->maxartsize)) {
 	  /* avoid buffer extention for ever */
 	  movedata = true;
 	} else {
@@ -1103,9 +1104,9 @@ NCproc(CHANNEL *cp)
     if (cp->State == CSwritegoodbye || cp->Type == CTfree)
       break;
     if (Tracing || cp->Tracing)
-      syslog(L_TRACE, "%s NCproc state=%d Start=%d Next=%d Used=%lu",
-	CHANname(cp), cp->State, cp->Start, cp->Next,
-        (unsigned long) bp->used);
+      syslog(L_TRACE, "%s NCproc state=%d Start=%lu Next=%lu Used=%lu",
+	CHANname(cp), cp->State, (unsigned long) cp->Start,
+        (unsigned long) cp->Next, (unsigned long) bp->used);
 
     if (movedata) { /* move data rather than extend buffer */
       TMRstart(TMR_DATAMOVE);
